refactor(server): split team.ts — extract 4 player handlers (S27.8.24)
Third slice on `apps/server/src/routes/team.ts`. Extract the 4
Player CRUD handlers into a dedicated module
`routes/team-player-handlers.ts` (578 lines).

Endpoints covered:
- `POST /:id/players` — handleAddTeamPlayer (budget + slots +
  unique number + position max)
- `DELETE /:id/players/:playerId` — handleDeleteTeamPlayer (funeral
  L2.B.6 exception)
- `PUT /:id/players/:playerId/skills` — handleUpdatePlayerSkills
  (primary/secondary chosen/random advancement, 6-cap, SPP cost,
  category access)
- `GET /:id/available-positions` — handleListAvailablePositions
  (counts + canAdd flag)

All thematically cohesive (Player CRUD on team roster). Re-export
from `team.ts` to preserve the public API consumed by
`team.test.ts`.

Leaf dependencies only (`prisma`, `sendError/sendSuccess`,
`updateTeamValues`, types and helpers from `@bb/game-engine`,
`getRosterFromDb`, `serverLog`, dynamic import `getPlayerCost`).
No cycle.

Cleanup of orphaned imports in `team.ts` (6 removed):
`getNextAdvancementPspCost`, `getPositionCategoryAccess`,
`SKILLS_BY_SLUG`, `SKILLS_DEFINITIONS`, `AdvancementType`,
`PlayerAdvancement`.

`team.ts` shrinks 1945 -> 1485 lines (-460 net lines, cumulative
-929 vs origin 2414 since S27.8.22 = -38%).

7 new smoke tests `team-player-handlers.test.ts`:
- exports x2 (named exports + re-export from team.ts)
- 404 gates on team not found x3
- 400 gates on missing body fields x2

- 1814/1814 server tests pass (was 1807, +7 new)
- typecheck clean
